fix.sh:optimization code

+++ fix.sh
@@ -3,7 +3,7 @@
 # Author: Carbon (ecrasy@gmail.com)
 # Description: feel free to use
 # Created Time: 2022-07-30 04:57:44 UTC
-# Modified Time: 2025-02-22 05:47:40 UTC
+# Modified Time: 2025-02-22 05:58:38 UTC
 #########################################################################
 
 
@@ -59,13 +59,13 @@ echo "Touch coremark log file to fix uhttpd error!!!"
 
 # Try latest dnsmasq
 tmp_ver=$(grep -m1 'PKG_UPSTREAM_VERSION:=' $GITHUB_WORKSPACE/data/dnsmasq/Makefile)
-tmp_ver=$(grep -m1 'PKG_RELEASE:=' ${dnsmasq_path}/Makefile)
-dnsmasq_data_ver="${tmp_ver##*=}.${tmp_ver##*=}"
+tmp_pkg=$(grep -m1 'PKG_RELEASE:=' $GITHUB_WORKSPACE/data/dnsmasq/Makefile)
+dnsmasq_data_ver="${tmp_ver##*=}.${tmp_pkg##*=}"
 if [ -n "${dnsmasq_data_ver}" ]; then
     dnsmasq_path="package/network/services/dnsmasq"
     tmp_ver=$(grep -m1 'PKG_UPSTREAM_VERSION:=' ${dnsmasq_path}/Makefile)
-    tmp_ver=$(grep -m1 'PKG_RELEASE:=' ${dnsmasq_path}/Makefile)
-    dnsmasq_repo_ver="${tmp_ver##*=}.${tmp_ver##*=}"
+    tmp_pkg=$(grep -m1 'PKG_RELEASE:=' ${dnsmasq_path}/Makefile)
+    dnsmasq_repo_ver="${tmp_ver##*=}.${tmp_pkg##*=}"
     cr=$(version_comp "${dnsmasq_repo_ver}" "${dnsmasq_data_ver}")
     if [ "$cr" == "<" ]; then
         rm -rf $dnsmasq_path
@@ -79,13 +79,13 @@ fi
 
 # Try latest golang
 tmp_ver=$(grep -m1 'GO_VERSION_MAJOR_MINOR:=' $GITHUB_WORKSPACE/data/golang/golang/Makefile)
-data_pkg=$(grep -m1 'GO_VERSION_PATCH:=' $GITHUB_WORKSPACE/data/golang/golang/Makefile)
-golang_data_ver="${tmp_ver##*=}.${data_pkg##*=}"
+tmp_pkg=$(grep -m1 'GO_VERSION_PATCH:=' $GITHUB_WORKSPACE/data/golang/golang/Makefile)
+golang_data_ver="${tmp_ver##*=}.${tmp_pkg##*=}"
 if [ -n "${golang_data_ver}" ]; then
     golang_path="feeds/packages/lang/golang"
     tmp_ver=$(grep -m1 'GO_VERSION_MAJOR_MINOR:=' ${golang_path}/golang/Makefile)
-    repo_pkg=$(grep -m1 'GO_VERSION_PATCH:=' ${golang_path}/golang/Makefile)
-    golang_repo_ver="${tmp_ver##*=}.${repo_pkg##*=}"
+    tmp_pkg=$(grep -m1 'GO_VERSION_PATCH:=' ${golang_path}/golang/Makefile)
+    golang_repo_ver="${tmp_ver##*=}.${tmp_pkg##*=}"
     cr=$(version_comp "${golang_repo_ver}" "${golang_data_ver}")
     if [ "$cr" == "<" ]; then
         rm -rf $golang_path
@@ -101,15 +101,17 @@ tmp_ver=$(grep -m1 'PKG_VERSION:=' ${GITHUB_WORKSPACE}/data/v2ray-core/Makefile)
 v2ray_data_ver="${tmp_ver##*=}"
 if [ -n "${v2ray_data_ver}" ]; then
     v2ray_path="feeds/packages/net/v2ray-core"
-    tmp_ver=$(grep -m1 'PKG_VERSION:=' ${v2ray_path}/Makefile)
-    v2ray_repo_ver="${tmp_ver##*=}"
-    cr=$(version_comp "${v2ray_repo_ver}" "${v2ray_data_ver}")
-    if [ "$cr" == "<" ]; then
-        rm -rf $v2ray_path
-        cp -r $GITHUB_WORKSPACE/data/v2ray-core ${v2ray_path}
-        echo "Upgrade v2ray-core from ${v2ray_repo_ver} to ${v2ray_data_ver}"
-    else
-        echo "V2ray-core no change need to make: ${v2ray_repo_ver}"
+    if [ -d "${v2ray_path}" ]; then
+        tmp_ver=$(grep -m1 'PKG_VERSION:=' ${v2ray_path}/Makefile)
+        v2ray_repo_ver="${tmp_ver##*=}"
+        cr=$(version_comp "${v2ray_repo_ver}" "${v2ray_data_ver}")
+        if [ "$cr" == "<" ]; then
+            rm -rf $v2ray_path
+            cp -r $GITHUB_WORKSPACE/data/v2ray-core ${v2ray_path}
+            echo "Upgrade v2ray-core from ${v2ray_repo_ver} to ${v2ray_data_ver}"
+        else
+            echo "V2ray-core no change need to make: ${v2ray_repo_ver}"
+        fi
     fi
 fi
